// src/types/ProgressInfo.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ProgressInfo = { current: number, total: number, percentage: number, unit: string, elapsedTime: number, estimatedRemaining: number | null, itemsPerSecond: number, status: string, alternativeCurrent: number, alternativeTotal: number, alternativeUnit: string, };